Major view refactor.
Done lots of tidying up, removing and generally simplifying the code.
There was lots of extra stuff left over from pulling code out of
Django. Backed out the attempt at a change_list support for now as
it seems almost impossible to make it work on both 1.3 and trunk
without re-writing large chuncks of the admin (Particularly due
to the signature change in ChangeList's __init__).

Bumped up the version to 0.1.0 final, seems like a decent place to
stop for the first realise and everything is working quite nicely
even with a fairly limited set of features.